// src/types/certificateElement.types.ts
export enum ConfigTypes {
  CATEGORY = "CATEGORY",
  UE_CHECKBOX_BOOLEAN = "UE_CHECKBOX_BOOLEAN",
  UE_CHECKBOX_CODE = "UE_CHECKBOX_CODE",
  UE_CHECKBOX_DATE = "UE_CHECKBOX_DATE",
  UE_CHECKBOX_MULTIPLE_CODE = "UE_CHECKBOX_MULTIPLE_CODE",
  UE_CHECKBOX_MULTIPLE_DATE = "UE_CHECKBOX_MULTIPLE_DATE",
  UE_CHECKBOX_MULTIPLE_DATE_RANGE = "UE_CHECKBOX_MULTIPLE_DATE_RANGE",
  UE_DATE = "UE_DATE",
  UE_DATE_RANGE = "UE_DATE_RANGE",
  UE_DIAGNOSES = "UE_DIAGNOSES",
  UE_DROPDOWN = "UE_DROPDOWN",
  UE_RADIO_BOOLEAN = "UE_RADIO_BOOLEAN",
  UE_RADIO_CODE = "UE_RADIO_CODE",
  UE_RADIO_MULTIPLE_CODE = "UE_RADIO_MULTIPLE_CODE",
  UE_RADIO_MULTIPLE_CODE_OPTIONAL_DROPDOWN = "UE_RADIO_MULTIPLE_CODE_OPTIONAL_DROPDOWN",
  UE_SICK_LEAVE_PERIOD = "UE_SICK_LEAVE_PERIOD",
  UE_TEXTAREA = "UE_TEXTAREA",
  UE_ICF = "UE_ICF",
  UE_UNCERTAIN_DATE = "UE_UNCERTAIN_DATE",
  UE_TEXTFIELD = "UE_TEXTFIELD",
  UE_TYPE_AHEAD = "UE_TYPE_AHEAD",
  UE_MESSAGE = "UE_MESSAGE",
  UE_HEADER = "UE_HEADER",
  UE_MEDICAL_INVESTIGATION = "UE_MEDICAL_INVESTIGATION",
  UE_CAUSE_OF_DEATH = "UE_CAUSE_OF_DEATH",
  UE_CAUSE_OF_DEATH_LIST = "UE_CAUSE_OF_DEATH_LIST",
  UE_VISUAL_ACUITY = "UE_VISUAL_ACUITY",
  UE_VIEW_TEXT = "UE_VIEW_TEXT",
  UE_VIEW_LIST = "UE_VIEW_LIST",
  UE_VIEW_TABLE = "UE_VIEW_TABLE",
  UE_YEAR = "UE_YEAR",
  UE_INTEGER = "UE_INTEGER",
}